Rename pkg/client_data -> pkg/_pub_shared[/data]
