+++ .github/workflows/dart.yml
@@ -37,16 +37,16 @@ jobs:
       - name: mono_repo self validate
         run: dart pub global run mono_repo generate --validate
   job_002:
-    name: "smoke_test; PKGS: app, pkg/_popularity, pkg/api_builder, pkg/client_data, pkg/fake_gcloud, pkg/pub_dartdoc, pkg/pub_dartdoc_data, pkg/pub_package_reader, pkg/pub_validations, pkg/web_app, pkg/web_css; `dart format --output=none --set-exit-if-changed .`, `dart analyze --fatal-infos  .`"
+    name: "smoke_test; PKGS: app, pkg/_popularity, pkg/_pub_shared, pkg/api_builder, pkg/fake_gcloud, pkg/pub_dartdoc, pkg/pub_dartdoc_data, pkg/pub_package_reader, pkg/pub_validations, pkg/web_app, pkg/web_css; `dart format --output=none --set-exit-if-changed .`, `dart analyze --fatal-infos  .`"
     runs-on: ubuntu-latest
     steps:
       - name: Cache Pub hosted dependencies
         uses: actions/cache@v2.1.7
         with:
           path: "~/.pub-cache/hosted"
-          key: "os:ubuntu-latest;pub-cache-hosted;sdk:2.15.1;packages:app-pkg/_popularity-pkg/api_builder-pkg/client_data-pkg/fake_gcloud-pkg/pub_dartdoc-pkg/pub_dartdoc_data-pkg/pub_package_reader-pkg/pub_validations-pkg/web_app-pkg/web_css;commands:format-analyze_0"
+          key: "os:ubuntu-latest;pub-cache-hosted;sdk:2.15.1;packages:app-pkg/_popularity-pkg/_pub_shared-pkg/api_builder-pkg/fake_gcloud-pkg/pub_dartdoc-pkg/pub_dartdoc_data-pkg/pub_package_reader-pkg/pub_validations-pkg/web_app-pkg/web_css;commands:format-analyze_0"
           restore-keys: |
-            os:ubuntu-latest;pub-cache-hosted;sdk:2.15.1;packages:app-pkg/_popularity-pkg/api_builder-pkg/client_data-pkg/fake_gcloud-pkg/pub_dartdoc-pkg/pub_dartdoc_data-pkg/pub_package_reader-pkg/pub_validations-pkg/web_app-pkg/web_css
+            os:ubuntu-latest;pub-cache-hosted;sdk:2.15.1;packages:app-pkg/_popularity-pkg/_pub_shared-pkg/api_builder-pkg/fake_gcloud-pkg/pub_dartdoc-pkg/pub_dartdoc_data-pkg/pub_package_reader-pkg/pub_validations-pkg/web_app-pkg/web_css
             os:ubuntu-latest;pub-cache-hosted;sdk:2.15.1
             os:ubuntu-latest;pub-cache-hosted
             os:ubuntu-latest
@@ -81,6 +81,19 @@ jobs:
         if: "always() && steps.pkg__popularity_pub_get.conclusion == 'success'"
         working-directory: pkg/_popularity
         run: dart analyze --fatal-infos  .
+      - id: pkg__pub_shared_pub_get
+        name: pkg/_pub_shared; dart pub get
+        if: "always() && steps.checkout.conclusion == 'success'"
+        working-directory: pkg/_pub_shared
+        run: dart pub get
+      - name: "pkg/_pub_shared; dart format --output=none --set-exit-if-changed ."
+        if: "always() && steps.pkg__pub_shared_pub_get.conclusion == 'success'"
+        working-directory: pkg/_pub_shared
+        run: "dart format --output=none --set-exit-if-changed ."
+      - name: "pkg/_pub_shared; dart analyze --fatal-infos  ."
+        if: "always() && steps.pkg__pub_shared_pub_get.conclusion == 'success'"
+        working-directory: pkg/_pub_shared
+        run: dart analyze --fatal-infos  .
       - id: pkg_api_builder_pub_get
         name: pkg/api_builder; dart pub get
         if: "always() && steps.checkout.conclusion == 'success'"
@@ -93,19 +106,6 @@ jobs:
       - name: "pkg/api_builder; dart analyze --fatal-infos  ."
         if: "always() && steps.pkg_api_builder_pub_get.conclusion == 'success'"
         working-directory: pkg/api_builder
-        run: dart analyze --fatal-infos  .
-      - id: pkg_client_data_pub_get
-        name: pkg/client_data; dart pub get
-        if: "always() && steps.checkout.conclusion == 'success'"
-        working-directory: pkg/client_data
-        run: dart pub get
-      - name: "pkg/client_data; dart format --output=none --set-exit-if-changed ."
-        if: "always() && steps.pkg_client_data_pub_get.conclusion == 'success'"
-        working-directory: pkg/client_data
-        run: "dart format --output=none --set-exit-if-changed ."
-      - name: "pkg/client_data; dart analyze --fatal-infos  ."
-        if: "always() && steps.pkg_client_data_pub_get.conclusion == 'success'"
-        working-directory: pkg/client_data
         run: dart analyze --fatal-infos  .
       - id: pkg_fake_gcloud_pub_get
         name: pkg/fake_gcloud; dart pub get
@@ -617,16 +617,16 @@ jobs:
       - job_005
       - job_006
   job_015:
-    name: "unit_test; PKG: pkg/client_data; `dart test --run-skipped`"
+    name: "unit_test; PKG: pkg/_pub_shared; `dart test --run-skipped`"
     runs-on: ubuntu-latest
     steps:
       - name: Cache Pub hosted dependencies
         uses: actions/cache@v2.1.7
         with:
           path: "~/.pub-cache/hosted"
-          key: "os:ubuntu-latest;pub-cache-hosted;sdk:2.15.1;packages:pkg/client_data;commands:test_07"
+          key: "os:ubuntu-latest;pub-cache-hosted;sdk:2.15.1;packages:pkg/_pub_shared;commands:test_07"
           restore-keys: |
-            os:ubuntu-latest;pub-cache-hosted;sdk:2.15.1;packages:pkg/client_data
+            os:ubuntu-latest;pub-cache-hosted;sdk:2.15.1;packages:pkg/_pub_shared
             os:ubuntu-latest;pub-cache-hosted;sdk:2.15.1
             os:ubuntu-latest;pub-cache-hosted
             os:ubuntu-latest
@@ -635,14 +635,14 @@ jobs:
           sdk: "2.15.1"
       - id: checkout
         uses: actions/checkout@v2.4.0
-      - id: pkg_client_data_pub_get
-        name: pkg/client_data; dart pub get
+      - id: pkg__pub_shared_pub_get
+        name: pkg/_pub_shared; dart pub get
         if: "always() && steps.checkout.conclusion == 'success'"
-        working-directory: pkg/client_data
+        working-directory: pkg/_pub_shared
         run: dart pub get
-      - name: "pkg/client_data; dart test --run-skipped"
-        if: "always() && steps.pkg_client_data_pub_get.conclusion == 'success'"
-        working-directory: pkg/client_data
+      - name: "pkg/_pub_shared; dart test --run-skipped"
+        if: "always() && steps.pkg__pub_shared_pub_get.conclusion == 'success'"
+        working-directory: pkg/_pub_shared
         run: dart test --run-skipped
     needs:
       - job_001
