+++ app/lib/frontend/handlers/pubapi.client.dart
@@ -4,11 +4,11 @@
 // ClientLibraryGenerator
 // **************************************************************************
 
+import 'package:_pub_shared/data/account_api.dart' as _i4;
+import 'package:_pub_shared/data/admin_api.dart' as _i6;
+import 'package:_pub_shared/data/package_api.dart' as _i3;
+import 'package:_pub_shared/data/publisher_api.dart' as _i5;
 import 'package:api_builder/_client_utils.dart' as _i2;
-import 'package:client_data/account_api.dart' as _i4;
-import 'package:client_data/admin_api.dart' as _i6;
-import 'package:client_data/package_api.dart' as _i3;
-import 'package:client_data/publisher_api.dart' as _i5;
 import 'package:http/http.dart' as _i1;
 
 export 'package:api_builder/_client_utils.dart' show RequestException;
