+++ app/lib/frontend/templates/package_versions.dart
@@ -2,7 +2,7 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import 'package:client_data/package_api.dart';
+import 'package:_pub_shared/data/package_api.dart';
 import 'package:collection/collection.dart' show IterableExtension;
 import 'package:pub_semver/pub_semver.dart';
