+++ app/lib/frontend/templates/publisher.dart
@@ -2,8 +2,8 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import 'package:client_data/page_data.dart';
-import 'package:client_data/publisher_api.dart' as api;
+import 'package:_pub_shared/data/page_data.dart';
+import 'package:_pub_shared/data/publisher_api.dart' as api;
 
 import '../../audit/models.dart';
 import '../../frontend/templates/views/account/activity_log_table.dart';
