+++ app/lib/frontend/templates/views/pkg/versions/version_row.dart
@@ -2,7 +2,7 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import 'package:client_data/package_api.dart';
+import 'package:_pub_shared/data/package_api.dart';
 
 import '../../../../../package/model_properties.dart';
 import '../../../../../shared/urls.dart' as urls;
