+++ app/test/admin/api_test.dart
@@ -4,8 +4,8 @@
 
 import 'dart:convert';
 
-import 'package:client_data/admin_api.dart';
-import 'package:client_data/package_api.dart';
+import 'package:_pub_shared/data/admin_api.dart';
+import 'package:_pub_shared/data/package_api.dart';
 import 'package:clock/clock.dart';
 import 'package:gcloud/db.dart';
 import 'package:pub_dev/account/backend.dart';
