+++ app/test/package/backend_test.dart
@@ -7,7 +7,7 @@ library pub_dartlang_org.backend_test;
 
 import 'dart:async';
 
-import 'package:client_data/package_api.dart';
+import 'package:_pub_shared/data/package_api.dart';
 import 'package:gcloud/db.dart';
 import 'package:pub_dev/account/backend.dart';
 import 'package:pub_dev/audit/backend.dart';
