+++ pkg/_pub_shared/build.yaml
@@ -5,8 +5,8 @@ targets:
     sources:
       - '$package$'
       - 'lib/$lib$'
-      - 'lib/admin_api.dart'
-      - 'lib/account_api.dart'
-      - 'lib/package_api.dart'
-      - 'lib/page_data.dart'
-      - 'lib/publisher_api.dart'
+      - 'lib/data/account_api.dart'
+      - 'lib/data/admin_api.dart'
+      - 'lib/data/package_api.dart'
+      - 'lib/data/page_data.dart'
+      - 'lib/data/publisher_api.dart'
