+++ pkg/_pub_shared/lib/data/package_api.dart
