+++ pkg/_pub_shared/lib/data/publisher_api.dart
