+++ pkg/_pub_shared/mono_pkg.yaml
